Added generic staging
To make the simulation more extensable, I have made the staging of diseases
follow a float64 adjacency matrix, with each row being the current stage and
each column being a next stage, with the elements of the matrix as the
probability of the transition. I have also added a default SEIR matrix for
simple studies.